requirements.txt
streamlit
langchain
PyPDF2
faiss-cpu
transformers
torch